FIX : 평균이 정수로 나오는 문제 수정
sum이 정수일 경우 실수로 캐스팅해야 한다!

+++ Lab03/src/Lab03_07.java
@@ -11,7 +11,7 @@ public class Lab03_07 {
 			sum += randomIntArray[i];
 			System.out.print(randomIntArray[i] + " ");
 		}
-		System.out.print("\n평균은 " + (sum / randomIntArray.length));
+		System.out.print("\n평균은 " + ((double)sum / randomIntArray.length));
 	}
 
 }
